backend/requirements.txt
anyio==4.11.0
attrs==25.4.0
backoff==2.2.1
base58==2.1.1
betterproto2==0.9.1
blinker==1.9.0
canoser==0.8.2
certifi==2025.10.5
click==8.3.0
dataclasses-json==0.6.7
deprecated==1.2.18
flask==3.1.2
flask-cors==6.0.1
gql==4.0.0
graphql-core==3.2.7
grpclib==0.4.8
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markupsafe==3.0.3
marshmallow==3.26.1
multidict==6.7.0
mypy-extensions==1.1.0
packaging==25.0
propcache==0.4.1
pysui==0.92.0
pysui-fastcrypto==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyyaml==6.0.3
referencing==0.37.0
rpds-py==0.28.0
six==1.17.0
sniffio==1.3.1
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-utils==0.1.0
websockets==15.0.1
werkzeug==3.1.3
wrapt==1.17.3
yarl==1.22.0
